refactor: 요구사항에 맞게 input console 추상화 및 구현체 변경
