+++ src/main/kotlin/calculator/interfaces/input/InputConsole.kt
@@ -1,0 +1,5 @@
+package calculator.interfaces.input
+
+interface InputConsole {
+    fun input(): String
+}
