+++ src/main/kotlin/calculator/interfaces/input/impl/SingleInputConsole.kt
@@ -1,0 +1,8 @@
+package calculator.interfaces.input.impl
+
+import calculator.interfaces.input.InputConsole
+
+object SingleInputConsole : InputConsole {
+
+    override fun input() = readln()
+}
